fix(tests): align persistence and integration tests with JSON migration
