+++ tests/test_index.py
@@ -104,108 +104,4 @@ def test_search_index_finds_most_similar(monkeypatch):
     assert results[0] == chunks[2]
 
 
-# ── pickle → JSON migration ─────────────────────────────────────────────────
 
-def test_save_index_uses_json(tmp_path, monkeypatch):
-    """save_index should write chunks as JSON, not pickle."""
-    monkeypatch.setattr(indexing, "PDF_CACHE_DIR", tmp_path)
-    monkeypatch.setattr(indexing, "INDEX_PATH", tmp_path / "index.faiss")
-    monkeypatch.setattr(indexing, "CHUNKS_PATH", tmp_path / "chunks.json")
-
-    chunks = _make_chunks(3)
-    vecs = np.random.randn(3, indexing.EMBED_DIM).astype(np.float32)
-    faiss.normalize_L2(vecs)
-    monkeypatch.setattr(indexing, "embed_texts", _make_embed_fn(vecs))
-    index = indexing.build_index(chunks)
-
-    indexing.save_index(index, chunks)
-
-    # Verify chunks.json is valid JSON (not pickle)
-    import json
-    chunks_file = tmp_path / "chunks.json"
-    assert chunks_file.exists()
-    with open(chunks_file, "r", encoding="utf-8") as f:
-        loaded = json.load(f)
-    assert len(loaded) == 3
-    assert loaded[0]["text"] == "chunk number 0"
-
-
-def test_load_precomputed_index_from_json(tmp_path, monkeypatch):
-    """load_precomputed_index should load chunks from JSON."""
-    import json
-    monkeypatch.setattr(indexing, "PDF_CACHE_DIR", tmp_path)
-    monkeypatch.setattr(indexing, "INDEX_PATH", tmp_path / "index.faiss")
-    monkeypatch.setattr(indexing, "CHUNKS_PATH", tmp_path / "chunks.json")
-
-    chunks = _make_chunks(3)
-    vecs = np.random.randn(3, indexing.EMBED_DIM).astype(np.float32)
-    faiss.normalize_L2(vecs)
-    monkeypatch.setattr(indexing, "embed_texts", _make_embed_fn(vecs))
-    index = indexing.build_index(chunks)
-
-    # Save as JSON
-    with open(tmp_path / "chunks.json", "w", encoding="utf-8") as f:
-        json.dump(chunks, f)
-    faiss.write_index(index, str(tmp_path / "index.faiss"))
-
-    # Load and verify
-    loaded_index, loaded_chunks = indexing.load_precomputed_index()
-    assert loaded_index.ntotal == 3
-    assert len(loaded_chunks) == 3
-    assert loaded_chunks[0]["text"] == "chunk number 0"
-
-
-def test_load_precomputed_index_deletes_legacy_pkl(tmp_path, monkeypatch):
-    """load_precomputed_index should delete legacy .pkl without loading it (security)."""
-    import pickle
-    monkeypatch.setattr(indexing, "PDF_CACHE_DIR", tmp_path)
-    monkeypatch.setattr(indexing, "INDEX_PATH", tmp_path / "index.faiss")
-    monkeypatch.setattr(indexing, "CHUNKS_PATH", tmp_path / "chunks.json")
-
-    chunks = _make_chunks(3)
-    vecs = np.random.randn(3, indexing.EMBED_DIM).astype(np.float32)
-    faiss.normalize_L2(vecs)
-    monkeypatch.setattr(indexing, "embed_texts", _make_embed_fn(vecs))
-    index = indexing.build_index(chunks)
-
-    # Save as legacy pickle (simulate pre-migration state)
-    pkl_path = tmp_path / "chunks.pkl"
-    with open(pkl_path, "wb") as f:
-        pickle.dump(chunks, f)
-    faiss.write_index(index, str(tmp_path / "index.faiss"))
-    # No chunks.json — triggers the legacy path
-
-    # JSON doesn't exist yet
-    assert not (tmp_path / "chunks.json").exists()
-
-    # Load should return None, None (not load the untrusted pickle)
-    loaded_index, loaded_chunks = indexing.load_precomputed_index()
-    assert loaded_index is None
-    assert loaded_chunks is None
-
-    # Pickle should be deleted, JSON should NOT be created from untrusted pickle
-    assert not pkl_path.exists()
-    assert not (tmp_path / "chunks.json").exists()
-
-
-def test_fetch_downloads_json_not_pkl(tmp_path, monkeypatch):
-    """_fetch_pdf_cache_if_missing should request chunks.json, not chunks.pkl."""
-    import urllib.request
-    monkeypatch.setattr(indexing, "PDF_CACHE_DIR", tmp_path)
-    monkeypatch.setattr(indexing, "INDEX_PATH", tmp_path / "index.faiss")
-    monkeypatch.setattr(indexing, "CHUNKS_PATH", tmp_path / "chunks.json")
-
-    requested_urls = []
-    def mock_urlretrieve(url, dest):
-        requested_urls.append(url)
-        # Create a dummy file so the function doesn't loop
-        open(dest, "w").write("[]")
-
-    monkeypatch.setattr(urllib.request, "urlretrieve", mock_urlretrieve)
-    monkeypatch.setenv("VEXILON_RAW_URL_BASE", "https://example.com")
-
-    indexing._fetch_pdf_cache_if_missing()
-
-    # Should have requested chunks.json, not chunks.pkl
-    assert any("chunks.json" in url for url in requested_urls)
-    assert not any("chunks.pkl" in url for url in requested_urls)
